src/requirements.txt
h5py==2.10.0
joblib==1.1.0
Keras-Preprocessing==1.1.2
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
Pillow==8.4.0
protobuf==3.19.4
scikit-learn==0.24.2
scipy==1.5.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tqdm==4.63.0
